fixed broken site

+++ Hearts of Gold/Hearts of Gold/Controllers/ItemsController.cs
@@ -12,6 +12,7 @@ using Microsoft.AspNet.Identity;
 
 namespace Hearts_of_Gold.Controllers
 {
+    [Authorize]
     public class ItemsController : Controller
     {
         private Hearts_Of_GoldEntities db = new Hearts_Of_GoldEntities();
@@ -26,10 +27,11 @@ namespace Hearts_of_Gold.Controllers
         }
 
         // GET: Items
+        [AllowAnonymous]
         public ActionResult Index()
         {
             var userId = ReturnUserId();
-
+            ViewBag.userID = HttpContext.User.Identity.GetUserId();
             var items = db.Items.Include(i => i.Donation_Categories).Include(i => i.Donation_Location).Include(i => i.User).ToList();
             var requests = db.Requests.Where(r => r.RequesterID == userId).ToList();
             var tuple = new Tuple<List<Item>,List<Request>>(items,requests); // combines the item and request into one object.
@@ -131,7 +133,8 @@ namespace Hearts_of_Gold.Controllers
 
             if (userId != userCompareId || itemId != item.ItemID)
             {
-                return RedirectToAction("Details", new { id = item.ItemID });
+                ViewData["error"] = "You can only edit your own item";
+                return View("Error");
             }
 
             if (ModelState.IsValid)
@@ -150,6 +153,7 @@ namespace Hearts_of_Gold.Controllers
         // GET: Items/Delete/5
         public ActionResult Delete(int? id)
         {
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -168,6 +172,15 @@ namespace Hearts_of_Gold.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Item item = db.Items.Find(id);
+            var userId = ReturnUserId();
+            var userCompareId = db.Items.Where(i => i.ItemID == item.ItemID).Select(i => i.UserID).FirstOrDefault();
+            var itemId = db.Items.Where(i => i.ItemID == item.ItemID).Select(i => i.ItemID).FirstOrDefault();
+
+            if (userId != userCompareId || itemId != item.ItemID)
+            {
+                ViewData["error"] = "You can only delete your own item";
+                return View("Error");
+            }
             db.Items.Remove(item);
             db.SaveChanges();
             return RedirectToAction("Index");
